ops/deploy/requirements.txt
apache-airflow-providers-celery
redis
mlflow==2.5.0
boto3==1.26.8
pandas
psycopg2-binary  # ✅ PostgreSQL 연결 라이브러리
scikit-learn
xgboost
pycaret
optuna
